slideNav component created
