+++ src/Components/App/Footer.css
@@ -1,31 +1,35 @@
-footer {
-  position: fixed;
-  bottom: 0;
-  left: 0;
-  height: 5vh;
-  width: 100%;
-  text-align: center;
-  font-size: 18px;
-  font-weight: 400;
-  border-top: 2px solid #41cdbc;
-  z-index: 20;
+.nav-footer {
+  margin-top: 7vh;
   display: grid;
-  grid-template-columns: 5fr 1fr 5fr;
-  column-gap: 3px;
+  grid-template-rows: 1vh 3vh 5vh;
+  row-gap: .5vh;
   align-items: center;
-  justify-items: left;
-}
-footer a {
-  color: var(--init-blue);
-  font-weight: 500;
-  justify-self: right;
+  height: 10vh;
 }
 
-footer img {
-  height: 32px;
-  justify-self: center;
+.nav-footer .nav-dev {
+  color: var(--init-charcoal);
+  font-size: 16px;
+  align-self: flex-end;
+  font-weight: 600;
 }
 
-.dark footer a {
-  color: var(--init-gold);
+.nav-footer svg {
+  color: var(--init-charcoal);
+  height: 1.2em;
+  width: 1.2em;
+}
+
+.nav-footer .pursuit-nav {
+  all:unset;
+  grid-row: 1;
+}
+
+.pursuit-nav img {
+  height: 4vh;
+  position: relative;
+  right: 1.5vw;
+}
+.nav-footer div {
+  grid-row: 3;
 }
